Move Cask’s test helpers to `test/`.
