+++ Library/Homebrew/cask/spec/cask/artifact/binary_spec.rb
@@ -1,9 +1,7 @@
 describe Hbc::Artifact::Binary do
   let(:cask) {
     Hbc::CaskLoader.load_from_file(TEST_FIXTURE_DIR/"cask/Casks/with-binary.rb").tap do |cask|
-      shutup do
-        InstallHelper.install_without_artifacts(cask)
-      end
+      InstallHelper.install_without_artifacts(cask)
     end
   }
   let(:expected_path) {
@@ -70,9 +68,7 @@ describe Hbc::Artifact::Binary do
   context "binary is inside an app package" do
     let(:cask) {
       Hbc::CaskLoader.load_from_file(TEST_FIXTURE_DIR/"cask/Casks/with-embedded-binary.rb").tap do |cask|
-        shutup do
-          InstallHelper.install_without_artifacts(cask)
-        end
+        InstallHelper.install_without_artifacts(cask)
       end
     }
