+++ Library/Homebrew/cask/test/cask/artifact/two_apps_correct_test.rb
@@ -15,7 +15,7 @@ describe Hbc::Artifact::App do
     let(:target_path_pro) { Hbc.appdir.join("Caffeine Pro.app") }
 
     before do
-      TestHelper.install_without_artifacts(cask)
+      InstallHelper.install_without_artifacts(cask)
     end
 
     it "installs both apps using the proper target directory" do
