+++ Library/Homebrew/cask/test/cask/artifact/uninstall_test.rb
@@ -9,7 +9,7 @@ describe Hbc::Artifact::Uninstall do
 
   before do
     shutup do
-      TestHelper.install_without_artifacts(cask)
+      InstallHelper.install_without_artifacts(cask)
     end
   end
