+++ Library/Homebrew/test/support/helper/cask/audit_matchers.rb
