+++ Library/Homebrew/test/support/helper/cask/cleanup.rb
