+++ Library/Homebrew/test/support/helper/cask/expectations_hash_helper.rb
