+++ Library/Homebrew/test/support/helper/cask/fake_dirs.rb
