+++ Library/Homebrew/test/support/helper/cask/fake_system_command.rb
