+++ Library/Homebrew/test/support/helper/cask/file_helper.rb
