+++ Library/Homebrew/test/support/helper/cask/kernel_at_exit_hacks.rb
