+++ Library/Homebrew/test/support/helper/cask/never_sudo_system_command.rb
