+++ Library/Homebrew/test/support/helper/cask/sha256_helper.rb
