+++ Library/Homebrew/test/support/helper/cask/shared_examples.rb
