+++ Library/Homebrew/test/support/helper/cask/shared_examples/dsl_base.rb
